Fixed hitung_luas_lingkaran() printing float

+++ hitung_luas.py
@@ -71,8 +71,8 @@ def hitung_luas_lingkaran():
 
     ruas_jari_jari = int(input("Masukan nilai ruas jari-jari\t: "))
 
-    hasil = 22/7 * jari_jari**2
-    hasil2 = 2 * 22/7 * ruas_jari_jari
+    hasil = int(22/7 * jari_jari**2)
+    hasil2 = int(2 * 22/7 * ruas_jari_jari)
 
     print("Luas Lingkaran\t\t:", hasil, "cm2")
     print("Keliling Lingkaran\t\t:", hasil2, "cm2")
